Support multiple PRs for same version in same repo
In larger IaC repos, having a single mega-bump-PR across all terraform states is
not always suitable. There may be many stakeholders and different CODEOWNERS,
or one broken terraform state where a failing workflow check is blocking all
others.

The action supports paths and excluding paths, but this does not play with the
checks for existing PR. By supplying an optional branch name suffix we can solve
this problem and allow multiple PRs for the same version in the same repository.